moved data wrapper to separate module
